Update cmds to run on CHAI single GPU BasicTrainer
This was due to four GPU setup not being stable on the CHAI cluster.
So we keep the batch_size the same, but up grad_accum_steps by 4x
to have a single GPU simulate the 64 batch_size of the previous
4 GPU setup.

Also FSDPTrainer was not even working with one GPU, let alone
four, so we revert back to BasicTrainer to get something working.